Module Tezos_dac_node_lib

module Certificate_streamers : sig ... end

Certificate_streamers handles the streams of certificate updates used to notify clients of a Coordinator node that a certificate has been updated.

module Configuration : sig ... end

module Daemon : sig ... end

module Data_streamer : sig ... end

Data_streamer is an in-memory data structure for handling pub-sub mechanism of streaming data from publishers to subscribers.

module Event : sig ... end

module Handler : sig ... end

module Node_context : sig ... end

module Page_store : sig ... end

module Pages_encoding : sig ... end

Library for encoding payloads of arbitrary size in formats that can be decoded by the Sc-rollup kernels.

module RPC_handlers : sig ... end

RPC_handlers is a module with handlers for DAC API endpoints.

module RPC_server : sig ... end

module Signature_manager : sig ... end

module Store : sig ... end

module Wallet_account : sig ... end

module Wallet_cctxt_helpers : sig ... end

Module that implements Account wallet functionalities.