Module Tezos_dac_node_lib.Certificate_streamers

Certificate_streamers handles the streams of certificate updates used to notify clients of a Coordinator node that a certificate has been updated.

type t

The type of Certificate_streamers. Values of type Certificate_streamers.t map root hashes to Certificate_repr.t Data_streamer.t.

val init : unit -> t

init () returns an empty map of certificate streams.

val handle_subscribe : Tezos_dac_lib.Dac_plugin.t -> t -> Tezos_dac_lib.Dac_plugin.raw_hash -> (Tezos_dac_lib.Certificate_repr.t Lwt_stream.t * Lwt_watcher.stopper) Tezos_base.TzPervasives.tzresult

handle_subscribe dac_plugin t root_hash creates a new watcher for root_hash in t, and returns the corresponding Lwt_stream and Lwt_watcher.stopper.

val push : Tezos_dac_lib.Dac_plugin.t -> t -> Tezos_dac_lib.Dac_plugin.raw_hash -> Tezos_dac_lib.Certificate_repr.t -> unit Tezos_base.TzPervasives.tzresult

push dac_plugin t root_hash certificate streams the updated certificate for root_hash.

val close : Tezos_dac_lib.Dac_plugin.t -> t -> Tezos_dac_lib.Dac_plugin.raw_hash -> bool Tezos_base.TzPervasives.tzresult

close dac_plugin t root_hash closes the certificate Data_streamer.t for root_hash. Returns true if the Data_streamer.t associated to root_hash was open before the function was invoked, false otherwise.