Module Tezos_dac_node_lib.Configuration

type host_and_port = {

host : string;

port : int;

}

host_and_port holds server configuration details of another dac node required to instantiate Dac_node_client.(#cctxt).

module Coordinator : sig ... end

Coordinator specific configuration.

module Committee_member : sig ... end

Committee_member specific configuration.

module Observer : sig ... end

Observer specific configuration.

type mode = private

| Coordinator of Coordinator.t

| Committee_member of Committee_member.t

| Observer of Observer.t

Mode specific fragment of a configuration.

type t = private {

data_dir : string;
(*
The path to the DAC node data directory.
*)

rpc_address : string;
(*
The address the DAC node listens to.
*)

rpc_port : int;
(*
The port the DAC node listens to.
*)

reveal_data_dir : string;
(*
The directory where the DAC node saves pages.
*)

mode : mode;
(*
Configuration parameters specific to the operating mode of the DAC.
*)

allow_v1_api : bool;
(*
Feature flag for registering V1 API endpoints.
*)

}

val mode_name : t -> string

mode_name t returns a string representation of the operating mode for the configuration t.

val make_coordinator : Tezos_crypto.Aggregate_signature.public_key list -> mode

make_coordinator committee_members creates a new coordinator configuration mode using the given committee_members public keys.

val make_committee_member : coordinator_rpc_address:Uri.t -> Tezos_crypto.Aggregate_signature.public_key_hash -> mode

make_committee_member ~coordinator_rpc_address committee_member_address creates a new committee member configuration with committee_member_address as the signer and coordinator_rpc_address as the coordinator.

val make_observer : committee_rpc_addresses:Uri.t list -> ?timeout:int -> Uri.t -> mode

make_observer ~committee_rpc_addresses ?timeout coordinator_rpc_address creates a new observer configuration that sets the DAC Committee rpc addresses to committee_rpc_addresses and Coordinator endpoint to coordinator_rpc_address.

val make : data_dir:string -> reveal_data_dir:string -> allow_v1_api:bool -> string -> int -> mode -> t

make ~data_dir ~reveal_data_dir ~allow_v1_api rpc_address rpc_port mode creates a configuration value from the specified parameters.

val filename : t -> string

filename config gets the path to config file

val data_dir_path : t -> string -> string

data_dir_path config subpath builds a subpath relatively to the config

val reveal_data_dir : t -> string

val default_data_dir : string

default_data_dir is the data directory that the DAC node will use, when one is not specified in the configuration: currently set to ${HOME}/.tezos-dac-node.

val default_reveal_data_dir : string

default_reveal_data_dir is the directory that the DAC node will use to store pages on disk. Currently set to ${HOME}/.tezos_rollup_node/wasm_2_0_0.

val default_rpc_address : string

default_rpc_address is the default address of the RPC server of the DAC node: currently set to "127.0.0.1"

val default_rpc_port : int

default_rpc_port is the default port of the RPC server of the DAC node: currently set to 10832.

val save : t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

save config writes config file in config.data_dir

val load : data_dir:string -> (t, Tezos_base.TzPervasives.Error_monad.tztrace) Stdlib.result Lwt.t

load ~data_dir config tries to load the configuration of the DAC node from data_dir.

val mode : t -> mode

mode config returns the mode specific configuration of config.