Module Tezos_dac_node_lib.Data_streamer

Data_streamer is an in-memory data structure for handling pub-sub mechanism of streaming data from publishers to subscribers.

type 'a t

'a t represents an instance of Data_streamer, where 'a is the type of the data that we stream.

val init : unit -> 'a t

Initializes an instance of Data_streamer.t.

val publish : 'a t -> 'a -> unit

publish streamer data publishes data to all attached subscribers of the streamer.

val handle_subscribe : 'a t -> 'a Lwt_stream.t * Lwt_watcher.stopper

handle_subscribe streamer returns a new stream of data for the subscriber to consume. An Lwt_watcher.stopper function is also returned for the subscriber to close the stream.

val close : 'a t -> unit

close streamer closes all the connections to streamer.