Module Tezos_dac_node_lib.Event

include module type of struct include Tezos_base.TzPervasives.Internal_event.Simple end

This module provides wrappers to declare events without having to declare a module and apply the Make functor. They are pretty-printed as <DOC> (<PARAMETER> = <VALUE>, ...) (unless there is no parameter at all, in which case they are printed as <DOC>). You may also use braces to inline some parameters in <DOC>. In the example below, hash is inlined.

For consistency, it is suggested that log messages do not start with a capital letter and do not end with a period. For instance, write ~msg: "started something important" instead of ~msg: "Started something important.". The reason is that messages that do not inline all of their parameters are followed by the remaining parameters in parentheses, and a period does not (arguably) look great in those cases. If it does not end with a period, it is not a sentence and thus capitalizing the first word does not make much sense either.

Declare events with one of the declare functions, for instance: let inject = Internal_event.Simple.declare_2 ~name: "inject" ~msg: "injected block {hash}" ("level", int) ("hash", string) You must declare events only once for a given name. Usually you should thus declare them as global variables.

There is one declare_n function for each number n of parameters. For instance, the above example uses declare_2 because it has two parameters: level and hash.

Each parameter has a default pretty-printer that you can override using the ?ppX parameters. For instance, to override the default pretty-printer of the second parameter, pass a ~pp2 argument. This allows to, for instance, print only the first element of a list, remove quotes for strings, decide how many decimal digits are printed for floats, etc.

The default pretty-printer behaves as follows:

for base types (booleans, integers, strings, floats), it uses OCaml syntax;

for strings which are longer than 64 characters, it prints only the first characters followed by an ellipsis [...];

for null or similar encodings, it prints N/A;

for unions of Null and another encoding e (this is the encoding of options), it prints null for Null and uses e to decide how to print the value otherwise;

for other constructed values (arrays, lists, objects, tuples, unions, and recursive types), it prints a placeholder like <list> or <obj> (see remark below).

Because constructed values such as lists and objects are printed as placeholders such as <list> and <obj> and thus do not give any information, those parameters are not automatically added in parentheses at the end of the messages. You can override the default pretty-printer of those parameters to override this behavior. In that case, values will be printed if your pretty-printer does not output an empty string. Note that you should also override the default pretty-printer of constructed types that you inline using braces, as they would be printed using the placeholder otherwise.

Then emit this event with some parameters like this: Internal_event.Simple.emit inject (42, "BL654654654645654654564") This event will be printed as: injected block BL654654654645654654564 (level = 42)

For all declare functions, the default value for level is Info.

type 'a t = 'a Tezos_event_logging.Internal_event.Simple.t

Event declarations where 'a is the type of the event parameters.

val emit : 'a t -> 'a -> unit Lwt.t

Emit an instance of an event.

val emit_at_top_level : 'a t -> 'a -> unit

Emit an instance of an event if called at toplevel

val emit__dont_wait__use_with_care : 'a t -> 'a -> unit

Emit an instance of an event but do not wait for completion. Only use if you really need to not wait for logging resolution before continuing. May cause increased memory usage and out-of-order log messages.

val declare_0 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> unit -> unit t

Declare an event with no parameters.

val declare_1 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> ?pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> 'a t

Declare an event with one parameter.

val declare_2 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> ?pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ('a * 'b) t

Declare an event with two parameters.

val declare_3 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> ?pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ('a * 'b * 'c) t

Declare an event with three parameters.

val declare_4 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> ?pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ('a * 'b * 'c * 'd) t

Declare an event with four parameters.

val declare_5 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> ?pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ?pp5:(Stdlib.Format.formatter -> 'e -> unit) -> (string * 'e Data_encoding.t) -> ('a * 'b * 'c * 'd * 'e) t

Declare an event with five parameters.

val declare_6 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> ?pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ?pp5:(Stdlib.Format.formatter -> 'e -> unit) -> (string * 'e Data_encoding.t) -> ?pp6:(Stdlib.Format.formatter -> 'f -> unit) -> (string * 'f Data_encoding.t) -> ('a * 'b * 'c * 'd * 'e * 'f) t

Declare an event with six parameters.

val declare_7 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> ?pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ?pp5:(Stdlib.Format.formatter -> 'e -> unit) -> (string * 'e Data_encoding.t) -> ?pp6:(Stdlib.Format.formatter -> 'f -> unit) -> (string * 'f Data_encoding.t) -> ?pp7:(Stdlib.Format.formatter -> 'g -> unit) -> (string * 'g Data_encoding.t) -> ('a * 'b * 'c * 'd * 'e * 'f * 'g) t

Declare an event with seven parameters.

val declare_8 : ?alternative_color:Tezos_event_logging.Internal_event.alternative_color -> ?section:string list -> name:string -> msg:string -> ?level:Tezos_event_logging.Internal_event.level -> ?pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ?pp5:(Stdlib.Format.formatter -> 'e -> unit) -> (string * 'e Data_encoding.t) -> ?pp6:(Stdlib.Format.formatter -> 'f -> unit) -> (string * 'f Data_encoding.t) -> ?pp7:(Stdlib.Format.formatter -> 'g -> unit) -> (string * 'g Data_encoding.t) -> ?pp8:(Stdlib.Format.formatter -> 'h -> unit) -> (string * 'h Data_encoding.t) -> ('a * 'b * 'c * 'd * 'e * 'f * 'g * 'h) t

Declare an event with eight parameters.

val section : string list

val starting_node : unit t

val rpc_server_started : unit t

val shutdown_node : int t

val store_is_ready : unit t

val operating_mode : string t

val rpc_server_is_ready : (string * int) t

val node_is_ready : unit t

val committee_keys_imported : unit t

val data_dir_not_found : string t

val layer1_node_new_head : (Tezos_base.TzPervasives.Block_hash.t * int32) t

val layer1_node_tracking_started : unit t

val protocol_plugin_resolved : string t

val protocol_plugin_not_resolved : (string * string) t

val daemon_error : Tezos_base.TzPervasives.Error_monad.tztrace t

val dac_threshold_not_reached : (int * int) t

val committee_member_not_in_wallet : Tezos_crypto.Aggregate_signature.Public_key_hash.t t

val committee_member_cannot_sign : Tezos_crypto.Signature.Public_key_hash.t t

val commit_member_no_public_key : Tezos_crypto.Signature.Public_key_hash.t t

val handle_new_subscription_to_hash_streamer : unit t

val subscribed_to_root_hashes_stream : unit t

val new_root_hash_received : string t

val new_subscription_to_certificate_updates : string t

val received_root_hash_processed : string t

val new_hash_pushed_to_data_streamer : string t

val processing_root_hash_failed : (string * Tezos_base.TzPervasives.Error_monad.tztrace) t

val new_signature_pushed_to_coordinator : Tezos_crypto.Aggregate_signature.t t

val no_committee_member_address : unit t

val cannot_retrieve_keys_from_address : Tezos_crypto.Aggregate_signature.Public_key_hash.t t

val fetched_missing_page : string t

val layer1_node_tracking_ended : unit t

layer1_node_tracking_ended is emitted when client loses connection to Tezos node.

val connection_lost : unit t

connection_lost is emitted when client looses connection to Coordinator's stream of root hashes.

val cannot_connect : (int * float * Tezos_base.TzPervasives.tztrace) t

cannot_connect event is emited when client tries to subscribe to the stream of root hashes, but fails.

val proto_short_hash_string : Tezos_base.TzPervasives.Protocol_hash.t -> string

val emit_protocol_plugin_resolved : Tezos_base.TzPervasives.Protocol_hash.t -> unit Lwt.t

val emit_protocol_plugin_not_resolved : Tezos_base.TzPervasives.Protocol_hash.t -> Tezos_base.TzPervasives.Protocol_hash.t -> unit Lwt.t

val emit_new_root_hash_received : Tezos_dac_lib.Dac_plugin.t -> Tezos_dac_lib__Dac_plugin.hash -> unit Lwt.t

val emit_new_subscription_to_certificate_updates : Tezos_dac_lib.Dac_plugin.hash -> unit Lwt.t

val emit_root_hash_pushed_to_data_streamer : Tezos_dac_lib.Dac_plugin.t -> Tezos_dac_lib__Dac_plugin.hash -> unit Lwt.t

val emit_received_root_hash_processed : Tezos_dac_lib.Dac_plugin.t -> Tezos_dac_lib__Dac_plugin.hash -> unit Lwt.t

val emit_processing_root_hash_failed : Tezos_dac_lib.Dac_plugin.t -> Tezos_dac_lib__Dac_plugin.hash -> Tezos_base.TzPervasives.Error_monad.tztrace -> unit Lwt.t

val emit_signature_pushed_to_coordinator : Tezos_crypto.Aggregate_signature.t -> unit Lwt.t

val emit_no_committee_member_address : unit -> unit Lwt.t

val emit_rpc_started : unit -> unit Lwt.t

val emit_cannot_retrieve_keys_from_address : Tezos_crypto.Aggregate_signature.Public_key_hash.t -> unit Lwt.t

val emit_fetched_missing_page : Tezos_dac_lib.Dac_plugin.hash -> unit Lwt.t

val emit_coordinators_connection_lost : unit -> unit Lwt.t

emit_cannot_connect_to_coordinator () emits connection_lost event.

val emit_cannot_connect_to_coordinator : count:int -> delay:float -> Tezos_base.TzPervasives.tztrace -> unit Lwt.t

emit_cannot_connect_to_coordinator ~count ~delay error emits cannot_connect event.

val emit_l1_tracking_ended : unit -> unit Lwt.t

emit_l1_tracking_ended () emits layer1_node_tracking_ended event.

val cannot_connect_to_tezos_node : count:int -> delay:float -> Tezos_base.TzPervasives.tztrace -> unit Lwt.t

emit_cannot_connect_to_tezos_node ~count ~delay error emits cannot_connect event.