Module Tezos_dac_node_lib.Handler

type t = unit Tezos_base.TzPervasives.tzresult Lwt.t * (unit -> unit)

A value of type t represent a handler that specifies a procedure that is invoked by the DAC daemon. In most cases, it never resolves unless the DAC node terminates, in which case the underlying Tezos_rpc__RPC_context.stopper is called.

val resolve_plugin_and_set_ready : Node_context.t -> t Tezos_base.TzPervasives.tzresult Lwt.t

Monitor heads and try resolve the DAC protocol plugin corresponding to the protocol of the targeted node.

val handlers : Node_context.t -> t Tezos_base.TzPervasives.tzresult Lwt.t list Tezos_base.TzPervasives.tzresult Lwt.t

handlers ctxt returns the handlers to be executed in the main loop of the daemon. The set of handlers is different according to the operating mode indicated in ctxt.mode. All operating modes have one handler for tracking new heads from the Layer 1. Additionally

If ctx.mode is Coordinator _, then no other handlers are present,

If ctxt.mode is Committee_member _, then the DAC node also has one handler for monitoring root hashes streamed by a Coordinator. Upon detecting a new root hash, the associated pages are downloaded from the Coordinator, after which the root hash is signed and the signature posted to the Coordinator,

If ctxt.mode is Observer _, then the DAC node also has one handler for monitoring root hashes streamed by a Coordinator. Upon detecting a new root hash, the associated pages are downloaded from the Coordinator. Differently from the case of Committee_member _, root hashes are not signed.