Module Tezos_dac_node_lib.Node_context

type dac_plugin_module = (module Tezos_dac_lib.Dac_plugin.T)

module Coordinator : sig ... end

Coordinator defines a partial Node_context.t that is available only to Coordinator nodes, and functions that can be used to operate on such mode-specific node contexts.

module Committee_member : sig ... end

Committee_member defines a partial Node_context.t that is available only to Committee_member nodes, and functions that can be used to operate on such mode specific partial node contexts.

module Observer : sig ... end

The type of an Observer specific partial Node_context.t.

type mode = private

| Coordinator of Coordinator.t

| Committee_member of Committee_member.t

| Observer of Observer.t

Operating mode specific fraction of a Node_context.t

type ready_ctxt = {

dac_plugin : dac_plugin_module;

}

A ready_ctx value contains globally needed information for a running dac node. It is available when the DAC plugin has been loaded.

type status =

| Ready of ready_ctxt

| Starting

The status of the dac node.

type t

A t value contains both the status and the dac node configuration. Its fields are available through accessors.

val init : Configuration.t -> Tezos_client_base.Client_context.full -> t Tezos_base.TzPervasives.tzresult Lwt.t

init config cctxt creates a t with a status set to Starting using the given dac node configuration config and tezos node client context cctxt.

exception Status_already_ready

Raised by set_ready when the status is already Ready _

val set_ready : t -> dac_plugin_module -> unit

set_ready ctxt dac_plugin_module updates in place the status value to Ready, and initializes the inner ready_ctxt value with the given parameters.

raises Status_already_ready
when the status is already Ready _

type Tezos_base.TzPervasives.error +=

| Node_not_ready

val get_ready : t -> ready_ctxt Tezos_base.TzPervasives.tzresult

get_ready ctxt extracts the ready_ctxt value from a context t. It propagates Node_not_ready if status is not ready yet. If called multiple times, it replaces current values for ready_ctxt with new one.

val get_status : t -> status

get_status ctxt returns the dac node status.

val get_mode : t -> mode

get_mode node_ctxt returns the operating mode specific fraction of a Node_context.t.

val get_tezos_node_cctxt : t -> Tezos_client_base.Client_context.full

get_tezos_node_cctxt ctxt returns the Tezos node's client context.

val get_dac_plugin : t -> Tezos_dac_lib.Dac_plugin.t Tezos_base.TzPervasives.tzresult

get_dac_plugin ctxt returns the Dac_plugin.t used in the node context. Fails with Node_not_ready when Node_context.status is not Ready.

val get_page_store : t -> Page_store.Filesystem.t

get_page_store ctxt returns the filesystem backed page store used by the Dac node to save and load pages.

val get_node_store : t -> 'a Tezos_layer2_store.Store_sigs.mode -> 'a Store.Irmin_store.t

get_node_store ctxt access_mode returns the Store.Irmin_store.t with access_mode used by Dac components.