Module Node_context.Committee_member

Committee_member defines a partial Node_context.t that is available only to Committee_member nodes, and functions that can be used to operate on such mode specific partial node contexts.

type t = private {

committee_member : Wallet_account.Committee_member.t;
(*
The Wallet_account.Committee_member wallet associated with the commitee_member managed by the DAC node.
*)

coordinator_cctxt : Tezos_dac_client_lib.Dac_node_client.cctxt;
(*
The Dac_node_client.cctxt used by the Committee_member node to send requests to a Coordinator node.
*)

}

The type of a Committee_member specific partial Node_context.t.

val secret_key_uri : t -> Tezos_client_base.Client_keys.sk_uri

secret_key_uri t returns the secret key URI associated with the committee member managed by the Committee_member node.