Module Node_context.Coordinator

Coordinator defines a partial Node_context.t that is available only to Coordinator nodes, and functions that can be used to operate on such mode-specific node contexts.

type t = private {

committee_members : Wallet_account.Coordinator.t list;
(*
The list of Wallet.Coordinator values associated with the Data Availability Committee members managed by the Coordinator node.
*)

hash_streamer : Tezos_dac_lib.Dac_plugin.raw_hash Data_streamer.t;
(*
The Dac_plugin.hash Data_streamer.t that the Coordinator node use to advertise root hashes to Committee_member and Observer nodes.
*)

certificate_streamers : Certificate_streamers.t;
(*
Certificate streamers to communicate to DAC clients when a certificate has been updated with a new signature.
*)

}

The type of a Coordinator specific partial Node_context.t.

val public_keys_opt : t -> Tezos_crypto.Aggregate_signature.public_key option list

public_keys_opt t returns the list of public keys associated with the data availability committee of t.

val committee_members : t -> Tezos_crypto.Aggregate_signature.public_key_hash list