Module Node_context.Observer

The type of an Observer specific partial Node_context.t.

type t = private {

coordinator_cctxt : Tezos_dac_client_lib.Dac_node_client.cctxt;
(*
The Dac_node_client.cctxt used by the Observer node to send requests to a Coordinator node.
*)

committee_cctxts : Tezos_dac_client_lib.Dac_node_client.cctxt list;
(*
The list of Dac_node_client.cctxt used by the Observer node to send requests to each Committee_member node respectively.
*)

timeout : int;
(*
Timeout in seconds for fetching a missing page from Committee_Member.
*)

}