Module Tezos_dac_node_lib.Page_store

type Tezos_base.TzPervasives.error +=

| Reveal_data_path_not_a_directory of string

| Cannot_create_reveal_data_dir of string

| Cannot_write_page_to_page_storage of {

hash : string;

content : bytes;

}

| Cannot_read_page_from_page_storage of string

| Incorrect_page_hash of {

expected : Tezos_dac_lib.Dac_plugin.raw_hash;

actual : Tezos_dac_lib.Dac_plugin.raw_hash;

}

| Cannot_fetch_remote_page_with_flooding_strategy of {

hash : Tezos_dac_lib.Dac_plugin.raw_hash;

}

module type S = sig ... end

S is the module type defining the backend required for persisting DAC pages data onto the page storage.

module Filesystem : S with type configuration = string

Filesystem is an implementation of the page store backed up by the local filesystem. The configuration is a string denoting the path where files will be saved. A page is saved as a file whose name is the hex encoded hash of its content.

type remote_configuration = {

cctxt : Tezos_dac_client_lib.Dac_node_client.cctxt;

page_store : Filesystem.t;

}

The configuration type for page stores that fetch pages remotely, when they are not available on the local page store. The parameter cctxt corresponds to the client context used to connect to the remote host for fetching pages.

module Remote : S with type configuration = remote_configuration

A Page_store implementation backed by the local filesystem, which uses a connection to a Dac node to retrieve pages that are not saved locally.

type remote_with_flooding_configuration = {

timeout : float;

cctxts : Tezos_dac_client_lib.Dac_node_client.cctxt list;

page_store : Filesystem.t;

}

A Page_store implementation backed by the local filesystem, which uses connections to a collection of Dac nodes to retrieve pages that are not saved locally. It retrieves pages by sending the same page request to each remote node and saves only 1 copy of the page locally.

It is typically used by the Observer node to retrieve missing pages from the DAC committee members.

module Remote_with_flooding : S with type configuration = remote_with_flooding_configuration

TODO: https://gitlab.com/tezos/tezos/-/issues/5672 Merge Remote_with_flooding into Remote