Module Tezos_dac_node_lib.Pages_encoding

Library for encoding payloads of arbitrary size in formats that can be decoded by the Sc-rollup kernels.

type Tezos_base.TzPervasives.error +=

| Payload_cannot_be_empty

| Cannot_serialize_page_payload

| Cannot_deserialize_page

| Non_positive_size_of_payload of int

| Merkle_tree_branching_factor_not_high_enough of int

| Cannot_combine_pages_data_of_different_type

| Hashes_page_repr_expected_single_element

module type CONFIG = sig ... end

CONFIG is a module for configuration of encoding scheme.

type version = int

module type VERSION = sig ... end

VERSION is a module for versioning Content and Hashes pages of a specific encoding.

module type Dac_codec = sig ... end

Dac_codec is a module for encoding a payload as a whole and returning the calculated root hash.

module type Buffered_dac_codec = sig ... end

Buffered_dac_codec icrementally constructs and serializes payload by aggregating messages one message at a time. add maintains partially constructed pages in a buffer and persist full pages to page_store. finalize is called to end the aggregation process, by flushing all the remaining data inside the buffer and calculating the resulting root hash.

module Merkle_tree : sig ... end

Encoding of DAC payload as a Merkle tree with an arbitrary branching factor greater or equal to 2. The serialization process works as follows: