Module Internal_for_tests.Make

Parameters

module B : Buffered_dac_codec

Signature

val serialize_payload : Tezos_dac_lib.Dac_plugin.t -> page_store:B.page_store -> bytes -> Tezos_dac_lib.Dac_plugin.hash Tezos_base.TzPervasives.tzresult Lwt.t

serialize_payload dac_plugin page_store payload serializes a payload into pages suitable to be read by a PVM's `reveal_preimage` and stores it into page_store. The serialization scheme is some hash-based encoding scheme, that produces a root hash representing the preimage.

val deserialize_payload : Tezos_dac_lib.Dac_plugin.t -> page_store:B.page_store -> Tezos_dac_lib.Dac_plugin.hash -> bytes Tezos_base.TzPervasives.tzresult Lwt.t

deserialize_payload dac_plugin page_store hash deserializes a payload from hash using some hash-based encoding scheme. Any payload serialized by serialize_payload can de deserialized from its root hash by deserialize_payload, that is, these functions are inverses of each other.