Parameter Make_buffered.S

type t

t represents an instance of a storage backened.

type configuration

configuration is the type of the configuration which can be used to initialize the store.

val init : configuration -> t

init configuration initializes a page store using the information provided in configuration.

val save : Tezos_dac_lib.Dac_plugin.t -> t -> hash:Tezos_dac_lib.Dac_plugin.hash -> content:bytes -> unit Tezos_base.TzPervasives.tzresult Lwt.t

save dac_plugin t ~hash ~content writes content of page onto storage backend t under given key.

When writing fails it returns a Cannot_write_page_to_page_storage error.

We require save to be atomic.

val mem : Tezos_dac_lib.Dac_plugin.t -> t -> Tezos_dac_lib.Dac_plugin.hash -> bool Tezos_base.TzPervasives.tzresult Lwt.t

mem dac_plugin t hash Checks whether there is an entry for hash in the page storage. It can fail with error Cannot_read_from_page_storage hash_as_string, if it was not possible to check that the page exists for any reason related to the page storage backend implementation.

val load : Tezos_dac_lib.Dac_plugin.t -> t -> Tezos_dac_lib.Dac_plugin.hash -> bytes Tezos_base.TzPervasives.tzresult Lwt.t

load dac_plugin t hash returns content of the storage backend t represented by a key. When reading fails it returns a Cannot_read_page_from_page_storage error.