Module type Pages_encoding.Buffered_dac_codec

Buffered_dac_codec icrementally constructs and serializes payload by aggregating messages one message at a time. add maintains partially constructed pages in a buffer and persist full pages to page_store. finalize is called to end the aggregation process, by flushing all the remaining data inside the buffer and calculating the resulting root hash.

type t

Buffer type.

type page_store

Page store type.

val empty : unit -> t

empty returns an empty buffer.

val add : Tezos_dac_lib.Dac_plugin.t -> page_store:page_store -> t -> bytes -> unit Tezos_base.TzPervasives.tzresult Lwt.t

add dac_plugin page_store buffer message adds a message to buffer. The buffer is serialized to page_store when it is full. Serialization logic is dependent on the encoding scheme.

val finalize : Tezos_dac_lib.Dac_plugin.t -> page_store:page_store -> t -> Tezos_dac_lib.Dac_plugin.hash Tezos_base.TzPervasives.tzresult Lwt.t

finalize dac_plugin page_store buffer serializes the buffer to page_store and returns a root hash that represents the final payload. The serialization logic is dependent on the encoding scheme. buffer is emptied after this call.

val deserialize_payload : Tezos_dac_lib.Dac_plugin.t -> page_store:page_store -> Tezos_dac_lib.Dac_plugin.hash -> bytes Tezos_base.TzPervasives.tzresult Lwt.t

deserialize_payload dac_plugin page_store hash deserializes a payload from hash using some hash-based encoding scheme. Any payload serialized by add + finalize can be deserialized by this function.