Module Signature_manager.Coordinator

Module that exposes signature operations necsesary when running in Configuration.Coordinator mode.

val handle_put_dac_member_signature : Node_context.Coordinator.t -> Tezos_dac_lib.Dac_plugin.t -> Tezos_layer2_store.Store_sigs.rw Store.Irmin_store.t -> Page_store.Filesystem.t -> Tezos_dac_lib.Signature_repr.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

handle_put_dac_member_signature ctx dac_plugin ro_store rw_store page_store dac_member_signature does the following procedure: 1. Checks that the root_hash provided inside Signature_repr.t is known. Fails if unknown 2. Checks that the dac_member_signature.signer_pkh is currently a Dac member. 3. Checks that the dac member has not yet signed. If already signed, then noop and return. 4. Otherwise: 1. Verify the signature against the root hash and signer's public key. 2. Add signature to Signature_store 3. Update the aggregate signature in Aggregate_signature_store