Module Tezos_dac_node_lib.Store

module Irmin_store : sig ... end

Irmin_store representation for Dac node.

module Signature_store : Tezos_layer2_store.Store_sigs.Nested_map with type 'a store = 'a Irmin_store.t and type primary_key = Tezos_dac_lib.Dac_plugin.hash and type secondary_key = Tezos_crypto.Aggregate_signature.public_key_hash and type value = Tezos_crypto.Aggregate_signature.signature

Signature_store is a nested map where primary_keys are root_hashes and secondary_keys are dac member public key hashes. Root hashes are strings instead of Dac_hash.t because we want to avoid runtime functorization of the module.

module Certificate_store : Tezos_layer2_store.Store_sigs.Map with type 'a store = 'a Irmin_store.t and type key = Tezos_dac_lib.Dac_plugin.raw_hash and type value = Tezos_dac_lib.Certificate_repr.t

Key-value store for Dac certificates where keys are hexified Dac_hash.t and values are Certificate_repr.t.