Parameter Make_append_only_map.V

type value

The type of the value being persisted on disk.

val name : string

name is a string for identifying values of this type. It is used by encoders of Map-like storage modules.

val encoding : value Tezos_base.TzPervasives.Data_encoding.t

encoding defines how values are serialized and deserialized when being written to or read from the underlying store.