Module Tezos_dac_node_lib.Wallet_account

module Coordinator : sig ... end

Module containing the definition of wallet and functions operating on wallets exclusive to Coordinator nodes.

module Committee_member : sig ... end

Module containing the definition of account wallet and functions operating on account wallets exclusive to Committee_member nodes.