Module Wallet_account.Committee_member

Module containing the definition of account wallet and functions operating on account wallets exclusive to Committee_member nodes.

type t = {

public_key_hash : Tezos_crypto.Aggregate_signature.public_key_hash;

secret_key_uri : Tezos_client_base.Client_keys.sk_uri;

}

The type of a wallet of a Committee_member node. It contains the public key hash of the single committee member associated with the Committee_member node, and its corresponding secret key URI. The wallet does not contain the public key of the committee member, as a Committee_member node never verifies the signatures of other committee members.

val of_committee_member_address : Tezos_crypto.Aggregate_signature.public_key_hash -> Tezos_client_base.Client_context.wallet -> t Tezos_base.TzPervasives.tzresult Lwt.t

of_committee_member_address pkh wallet_cctxt constructs a value of type Committee_member.t from the public key hash pkh, by using the wallet context wallet_cctxt .