Module Wallet_account.Coordinator

Module containing the definition of wallet and functions operating on wallets exclusive to Coordinator nodes.

type t = {

public_key_hash : Tezos_crypto.Aggregate_signature.public_key_hash;

public_key : Tezos_crypto.Aggregate_signature.public_key;

}

The type of an account wallet for Coordinator nodes. An account wallet contains the public key hash and a public key for each account wallet. Secret keys URIs are not included in Coordinator account wallets, as Coordinator nodes never sign any data.

val of_committee_member_public_key : Tezos_crypto.Aggregate_signature.public_key -> t Tezos_base.TzPervasives.tzresult Lwt.t

of_committee_member_public_key public_key constructs a value of type Coordinator.t from the public key public_key.