Module Tezos_dac_node_lib.Wallet_cctxt_helpers

Module that implements Account wallet functionalities.

val get_keys : Tezos_client_base.Client_context.wallet -> Tezos_crypto.Aggregate_signature.public_key_hash -> (Tezos_crypto.Aggregate_signature.public_key_hash * Tezos_crypto.Aggregate_signature.public_key option * Tezos_client_base.Client_keys.sk_uri option) Tezos_base.TzPervasives.tzresult Lwt.t

get_keys cctxt pkh returns the keys associated with the given public key hash pkh in the wallet context cctxt.

val get_public_key : Tezos_client_base.Client_context.wallet -> Tezos_crypto.Aggregate_signature.public_key_hash -> Tezos_crypto.Aggregate_signature.public_key option Tezos_base.TzPervasives.tzresult Lwt.t

get_public_key cctxt pkh returns the public key associated with the given pkh if it can be found in cctxt.

val can_verify : (Tezos_crypto.Aggregate_signature.public_key_hash * Tezos_crypto.Aggregate_signature.public_key option * Tezos_client_base.Client_keys.sk_uri option) -> bool

can_verify (pkh, pk_opt, sk_uri_opt) checks whether the public key pk_opt of an account is defined.

val can_sign : (Tezos_crypto.Aggregate_signature.public_key_hash * Tezos_crypto.Aggregate_signature.public_key option Tezos_base.TzPervasives.tzresult * Tezos_client_base.Client_keys.sk_uri option) -> bool

can_sign (pkh, pk_opt, sk_uri_opt) checks whether the secret key URI sk_uri_opt of an account is defined.